+++ app/src/main/scala/hstream/server/KafkaConfig.scala
@@ -13,15 +13,30 @@
 
 package kafka.server
 
+import java.io._
+import java.nio.file.{Path, Paths, StandardOpenOption}
 import java.util.Properties
-import org.apache.kafka.common.config.{AbstractConfig, ConfigDef, ConfigException, ConfigResource, SaslConfigs, SecurityConfig, SslClientAuth, SslConfigs, TopicConfig}
+import org.apache.kafka.common.config.{
+  AbstractConfig,
+  ConfigDef,
+  ConfigException,
+  ConfigResource,
+  SaslConfigs,
+  SecurityConfig,
+  SslClientAuth,
+  SslConfigs,
+  TopicConfig
+}
 import org.apache.kafka.common.security.auth.SecurityProtocol
 import org.apache.kafka.common.network.ListenerName
 
-import scala.collection.{Map, Seq, mutable}
+import scala.collection.{mutable, Map, Seq}
 import scala.annotation.nowarn
+import scala.sys.process._
+import org.yaml.snakeyaml.Yaml
 import kafka.cluster.EndPoint
 import kafka.utils.{CoreUtils, Logging}
+import scala.jdk.CollectionConverters._
 import kafka.utils.Implicits._
 
 object Defaults {
@@ -114,11 +129,26 @@ object KafkaConfig {
         MEDIUM,
         "$DefaultReplicationFactorDoc"
       )
-      .define(OffsetsTopicReplicationFactorProp, SHORT, Defaults.DefaultOffsetsTopicReplicationFactor, atLeast(1), HIGH, "$OffsetsTopicReplicationFactorDoc")
-      .define(GroupInitialRebalanceDelayMsProp, INT, Defaults.GroupInitialRebalanceDelayMs, MEDIUM, "$GroupInitialRebalanceDelayMsDoc")
+      .define(
+        OffsetsTopicReplicationFactorProp,
+        SHORT,
+        Defaults.DefaultOffsetsTopicReplicationFactor,
+        atLeast(1),
+        HIGH,
+        "$OffsetsTopicReplicationFactorDoc"
+      )
+      .define(
+        GroupInitialRebalanceDelayMsProp,
+        INT,
+        Defaults.GroupInitialRebalanceDelayMs,
+        MEDIUM,
+        "$GroupInitialRebalanceDelayMsDoc"
+      )
       .define(SaslKerberosServiceNameProp, STRING, null, MEDIUM, "$SaslKerberosServiceNameDoc")
 
-      /** ********* Fetch Configuration **************/
+      /**
+       * ********* Fetch Configuration *************
+       */
       .define(FetchMaxBytes, INT, Defaults.FetchMaxBytes, atLeast(1024), MEDIUM, "$FetchMaxBytesDoc")
 
     // TODO: KAFKA_ORIGINAL
@@ -203,7 +233,10 @@ class KafkaConfig private (
     properties.put(KafkaConfig.NumPartitionsProp, getInt(KafkaConfig.NumPartitionsProp))
     properties.put(KafkaConfig.DefaultReplicationFactorProp, getInt(KafkaConfig.DefaultReplicationFactorProp))
     properties.put(KafkaConfig.AutoCreateTopicsEnableProp, getBoolean(KafkaConfig.AutoCreateTopicsEnableProp))
-    properties.put(KafkaConfig.OffsetsTopicReplicationFactorProp, getShort(KafkaConfig.OffsetsTopicReplicationFactorProp))
+    properties.put(
+      KafkaConfig.OffsetsTopicReplicationFactorProp,
+      getShort(KafkaConfig.OffsetsTopicReplicationFactorProp)
+    )
     if (!props.containsKey(KafkaConfig.GroupInitialRebalanceDelayMsProp))
       properties.put(KafkaConfig.GroupInitialRebalanceDelayMsProp, "0")
     else
@@ -299,6 +332,174 @@ class KafkaConfig private (
           "Error parsing configuration property '%s': %s".format(propName, e.getMessage)
         )
     }
+  }
+
+}
+
+object TestConfig {
+
+  val currentTestTimeMillis = System.currentTimeMillis()
+
+  def parseConfigFile(
+      configFile: String,
+      startingIdNumber: Int,
+      endingIdNumber: Int,
+      testName: String,
+      unusedPorts: Seq[Int]
+  ): Seq[Properties] = {
+    val configs = readConfigFile(configFile)
+    val use = configs
+      .getOrElse("use", throw new IllegalArgumentException("use is required in config file"))
+      .asInstanceOf[String]
+    if (use == "broker_container") {
+      // Start hserver by docker
+      val brokerContainer = configs
+        .getOrElse(
+          "broker_container",
+          throw new IllegalArgumentException("broker_container is required in config file")
+        )
+        .asInstanceOf[java.util.Map[String, Object]]
+      return createBrokerConfigsFromBrokerContainer(
+        brokerContainer,
+        startingIdNumber,
+        endingIdNumber,
+        testName,
+        unusedPorts
+      )
+    } else if (use == "broker_connections") {
+      // Directly connect to brokers
+      val brokerConnections =
+        configs
+          .getOrElse(
+            "broker_connections",
+            throw new IllegalArgumentException("broker_connections is required in config file")
+          )
+          .asInstanceOf[java.util.List[java.util.Map[String, Object]]]
+      brokerConnections.asScala.map(_.asScala.toMap).map { config =>
+        val props = new Properties
+        config.foreach { case (k, v) => props.put(k, v.toString) }
+        props
+      }
+    } else {
+      throw new IllegalArgumentException("use must be one of [broker_container, broker_connections]")
+    }
+  }
+
+  private def createBrokerConfigsFromBrokerContainer(
+      brokerContainer: java.util.Map[String, Object],
+      startingIdNumber: Int,
+      endingIdNumber: Int,
+      testName: String,
+      unusedPorts: Seq[Int]
+  ): Seq[Properties] = {
+    val brokerConfig = brokerContainer.get("config").asInstanceOf[java.util.Map[String, Object]].asScala
+    val testingConfig: collection.mutable.Map[String, Object] =
+      brokerContainer
+        .get("testing_config")
+        .asInstanceOf[java.util.Map[String, Object]]
+        .asScala
+
+    val advertisedAddress = brokerConfig
+      .getOrElse(
+        "advertised.address",
+        throw new IllegalArgumentException("advertised.address is required in broker_container")
+      )
+      .asInstanceOf[String]
+    val basePort = testingConfig.get("base_port").asInstanceOf[Option[Int]]
+    val initMode = testingConfig.getOrElse("init_mode", "boot").asInstanceOf[String]
+    val commandTmpl = testingConfig
+      .getOrElse(
+        "command",
+        throw new IllegalArgumentException("command is required in testing_config")
+      )
+      .asInstanceOf[String]
+    // TODO: support rqlite
+    val metastorePort = testingConfig
+      .getOrElse("metastore_port", throw new IllegalArgumentException("metastore_port is required in testing_config"))
+      .asInstanceOf[Int]
+    val cleanContainerLog = testingConfig.getOrElse("container_logs_clean", false).asInstanceOf[Boolean]
+
+    var seedNodes = ""
+
+    // generate
+    val props = (startingIdNumber to endingIdNumber).zipWithIndex.map { case (nodeId, idx) =>
+      val prop = new Properties
+      val port = basePort match {
+        case None    => unusedPorts(idx * 2)
+        case Some(p) => p + idx * 2
+      }
+      val gossipPort = basePort match {
+        case None    => unusedPorts(idx * 2 + 1)
+        case Some(p) => p + idx * 2 + 1
+      }
+      if (initMode == "join") {
+        // we use a fixed port (the first gossipPort) for seed node when initMode is "join"
+        if (idx == 0) {
+          seedNodes = s"127.0.0.1:$gossipPort"
+        }
+      } else if (initMode == "boot") {
+        val sepr = if (idx == 0) "" else ","
+        seedNodes += s"${sepr}127.0.0.1:$gossipPort"
+      } else {
+        throw new IllegalArgumentException("init_mode must in [\"boot\", \"join\"]")
+      }
+      // broker config
+      prop.put("broker.id", nodeId.toString)
+      prop.put("port", port.toString)
+      prop.put("gossip.port", gossipPort.toString)
+      brokerConfig.foreach { case (k, v) => prop.put(k, v.toString) }
+      // testing config
+      val args = s"""--server-id $nodeId
+                --port $port --gossip-port $gossipPort
+                --advertised-address $advertisedAddress
+                --metastore-uri zk://127.0.0.1:$metastorePort
+                --store-config /data/store/logdevice.conf
+                """.stripMargin.linesIterator.mkString(" ").trim
+      (prop, args)
+    }
+    // The seedNodes is determined after all broker configs are created
+    return props.map {
+      case (prop, args) => {
+        val newArgs = s"$args --seed-nodes $seedNodes"
+        var newTestingConfig = testingConfig.clone()
+        newTestingConfig.update("command", commandTmpl.format(newArgs))
+        newTestingConfig.put("container_logs_dir", getContainerLogsDir(testName, cleanContainerLog))
+        prop.put("testing", newTestingConfig)
+        prop
+      }
+    }
+  }
+
+  private def readConfigFile(configFile: String): Map[String, Object] = {
+    val inputStream = new FileInputStream(configFile)
+    try {
+      val yaml = new Yaml()
+      yaml
+        .load(new InputStreamReader(inputStream, "UTF-8"))
+        .asInstanceOf[java.util.Map[String, Object]]
+        .asScala
+    } catch {
+      case e: Throwable => throw e
+    } finally {
+      inputStream.close()
+    }
+  }
+
+  private def formatTestNameAsFile(testName: String) = {
+    testName.replaceAll("""\(|\)|\s|\[|\]|=""", "_").replaceAll("(^_*)|(_*$)", "")
+  }
+
+  private def getContainerLogsDir(testName: String, cleanBefore: Boolean = false): Path = {
+    val testFilename = formatTestNameAsFile(testName)
+    val proj = sys.props.get("user.dir").getOrElse(".")
+    val containerLogsDir = s"$proj/build/reports/logs/$testFilename-$currentTestTimeMillis"
+    // FIXME: how about moving clean function into "QuorumTestHarness",
+    // Which means, in QuorumTestHarness, do "rm -rf $proj/build/reports/logs" once.
+    if (cleanBefore) {
+      // TODO: Safer way
+      s"bash -c \"rm -rf $proj/build/reports/logs/$testFilename-*\"".!
+    }
+    Paths.get(containerLogsDir)
   }
 
 }
